Crate ungrammar

Ungrammar – a DSL for specifying concrete syntax tree grammar.

Producing a parser is an explicit non-goal – it’s ok for this grammar to be ambiguous, non LL, non LR, etc.

See this introductory post for details.

Structs

Error

An error encountered when parsing a Grammar.

Grammar

An Ungrammar grammar.

Node

A node, like A = 'b' | 'c'.

NodeData

Data about a node.

Token

A token, denoted with single quotes, like '+' or 'struct'.

TokenData

Data about a token.

Enums

Rule

A production rule.

Functions

rust_grammar

Returns a Rust grammar.

Type Definitions

Result

A type alias for std’s Result with the Error as our error type.