ungrammar

Struct Grammar

pub struct Grammar { /* private fields */ }

An Ungrammar grammar.

Implementations

impl Grammar

pub fn iter(&self) -> impl Iterator<Item = Node> + '_

Returns an iterator over all nodes in the grammar.

pub fn tokens(&self) -> impl Iterator<Item = Token> + '_

Returns an iterator over all tokens in the grammar.